프로젝트 사이클에 따른 지표 할당을 하드코딩이 아닌 유형화 시킴 (PRE_DMA_G0 : 경영일반 / POST_DMA_DISCLOSURE: 이중중대성평가 이후 선정된 지표할당) #138
