+++ backend/src/apis/onboardingApproval.py
@@ -14,6 +14,7 @@ from src.services.onboardings.service import (
     getApprovalStatus,
     listApprovals,
     rejectApproval,
+    reviewApproval,
     submitApproval,
 )
 from src.utils.auth import get_token
@@ -30,12 +31,12 @@ onboardingApprovalRouter = APIRouter(
 @router.post(
     "/submit",
     response_model=OnboardingApprovalActionResponseDto,
-    summary="Submit PRE_DMA_G0 G0-02 inputs for approval",
+    summary="Submit onboarding metric inputs for approval",
 )
 @onboardingApprovalRouter.post(
     "/submit",
     response_model=OnboardingApprovalActionResponseDto,
-    summary="Submit PRE_DMA_G0 G0-02 inputs for approval",
+    summary="Submit onboarding metric inputs for approval",
 )
 async def submitApprovalRoute(
     request: OnboardingApprovalRequestDto,
@@ -80,21 +81,22 @@ async def listApprovalsRoute(
 @router.get(
     "/status",
     response_model=OnboardingApprovalStatusResponseDto,
-    summary="Get PRE_DMA_G0 G0-02 approval status",
+    summary="Get onboarding metric approval status",
 )
 @onboardingApprovalRouter.get(
     "/status",
     response_model=OnboardingApprovalStatusResponseDto,
-    summary="Get PRE_DMA_G0 G0-02 approval status",
+    summary="Get onboarding metric approval status",
 )
 async def getApprovalStatusRoute(
     companyId: int = Query(...),
     reportingYear: int = Query(...),
+    cycleType: str = Query(default="PRE_DMA_G0"),
     metricId: str = Query(default="G0-02"),
     userModel=Depends(get_token),
 ):
     try:
-        return getApprovalStatus(companyId, reportingYear, metricId, userModel)
+        return getApprovalStatus(companyId, reportingYear, metricId, userModel, cycleType)
     except PermissionError as e:
         raise HTTPException(status_code=403, detail=str(e))
     except ValueError as e:
@@ -106,12 +108,12 @@ async def getApprovalStatusRoute(
 @router.post(
     "/approve",
     response_model=OnboardingApprovalActionResponseDto,
-    summary="Approve PRE_DMA_G0 G0-02 inputs and promote to KPI facts",
+    summary="Approve onboarding metric inputs",
 )
 @onboardingApprovalRouter.post(
     "/approve",
     response_model=OnboardingApprovalActionResponseDto,
-    summary="Approve PRE_DMA_G0 G0-02 inputs and promote to KPI facts",
+    summary="Approve onboarding metric inputs",
 )
 async def approveApprovalRoute(
     request: OnboardingApprovalDecisionRequestDto,
@@ -128,14 +130,38 @@ async def approveApprovalRoute(
 
 
 @router.post(
+    "/review",
+    response_model=OnboardingApprovalActionResponseDto,
+    summary="Review onboarding metric inputs",
+)
+@onboardingApprovalRouter.post(
+    "/review",
+    response_model=OnboardingApprovalActionResponseDto,
+    summary="Review onboarding metric inputs",
+)
+async def reviewApprovalRoute(
+    request: OnboardingApprovalDecisionRequestDto,
+    userModel=Depends(get_token),
+):
+    try:
+        return reviewApproval(request, userModel)
+    except PermissionError as e:
+        raise HTTPException(status_code=403, detail=str(e))
+    except ValueError as e:
+        raise HTTPException(status_code=409, detail=str(e))
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
+
+
+@router.post(
     "/reject",
     response_model=OnboardingApprovalActionResponseDto,
-    summary="Reject PRE_DMA_G0 G0-02 inputs",
+    summary="Reject onboarding metric inputs",
 )
 @onboardingApprovalRouter.post(
     "/reject",
     response_model=OnboardingApprovalActionResponseDto,
-    summary="Reject PRE_DMA_G0 G0-02 inputs",
+    summary="Reject onboarding metric inputs",
 )
 async def rejectApprovalRoute(
     request: OnboardingApprovalDecisionRequestDto,
